requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
langchain
langchain-huggingface
passlib[bcrypt]
python-jose[cryptography]
